# Enhance performance for plain text syntax highlighting