--- a/changelog.d/20230910_163316_GitHub_Actions_plain-highlighting-performance.rst
+++ b/changelog.d/20230910_163316_GitHub_Actions_plain-highlighting-performance.rst
@@ -1,0 +1,7 @@
+.. _#1843:  https://github.com/fox0430/moe/pull/1843
+
+Changed
+.......
+
+- `1843`_ Enhanced performance for plain text syntax highlighting
+