… tests/tvisualmode.nim