Enhance performance for plain text syntax highlighting parsing speed when opening a 1GB and 13944700 lines plain file has become about 10x faster.

```
base64 /dev/urandom | head -c 1073741824  > file.txt
moe file.txt
```

| Before | After |
| --- | --- |
| 10 seconds, 879 milliseconds | 1 second, 672 milliseconds |